Refactor login page to add metadata
